fix(api): harden the policy catalog per review
- move the catalog read into lib/usage/policy-catalog with the
  SafeHandlerGenerator annotation so the Eden response stays typed
- checkout availability requires configured provider credentials and a
  non-empty provider reference
- add-on setup requires an existing hosted entitlement (409 otherwise)
- three-way price-field consistency, domain CHECKs for kind, interval,
  and visibility, strict ISO 4217 currency codes, seeds bounded at 100
